pyproject.toml
[tool.poetry]
name = "hirey"
version = "0.1.0"
description = ""
authors = ["Mamdouh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.2"}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
databases = {extras = ["postgresql"], version = "^0.9.0"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asgi-lifespan = "^2.1.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
pytest = "^8.2.0"
ruff = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
